// src/types/login-response.ts
export interface LoginResponseProps {
  cpf: string;
  message: string;
  nome: string;
  ref_id: string;
  status: string;
  token: string;
  apae_filiada: string;
  area_atuacao: string;
  coordenacao: string;
  data_nascimento: string;
  escolaridade: string;
  funcao: string;
  municipio: string;
  necessita_apoio: string;
  possui_deficiencia: string;
  sobrenome: string;
  tamanho_camiseta: string;
  telefone_contato: string;
  tipo_apoio: string;
  tipo_deficiencia: string;
  uf: string;
  whatsapp: string;
  email: string;
  presidente_apae: string;
  documentid: string;
  dataInscricao: string;
  cep: string;
  atividades: string[];
}